fix(early-dispatch): propagate dispatch fanin on gate open, gate ED behind ready queue
ED-released tasks bypass send_task, so they never bumped their successors'
dispatch_fanin. Every task ED won permanently killed its successors' ED
eligibility, capping coverage near 12% of tasks. Gate open, not stage, is the
correct propagation point: a staged task is still GATED and not yet runnable,
so accounting at stage time would let a successor read its predecessor as
ready before it actually is.

- fix: task_desc.duration, new_task() and executor_t.duration were uint16_t,
  silently truncating Qwen3 constants (DUR_GATE_PROJ 95700 -> 30164) and
  understating simulated makespan. Widened to uint32_t.
- dispatch: send_task prefers cores with both slots free for the type, then
  half-free cores, one task per core per round. Previously it ANDed both slot
  bitmaps, demanding a fully free core and needlessly stalling dispatch.
- dispatch: ED runs only when every normal ready queue is drained. Ready tasks
  have all predecessors complete and run as soon as they get a slot, while
  staged tasks may still wait on predecessors, so ready work claims slots first.
- early-dispatch: pick_stage_core prefers the core where a same-type
  predecessor ran, else any core with both slots free for that type. It never
  falls back to a busy core's second slot, which would erase ED's head start.
- metrics: add LAT_TRACE (default off, no runtime cost when disabled) splitting
  ready->run latency into ready->enq, enq->deq and deq->run segments.
- test: assert a busy core neither executes its waiting slot nor consumes that
  slot's doorbell, locking in the one-slot-at-a-time execution model.

Correctness only; ED coverage and makespan tuning stay open. Verified against
the HEAD baseline: unit test outcomes are identical across all 20 (target, ED)
combinations, and paged_attention (1920 tasks) plus qwen3 (864 tasks) each ran
20 times per ED setting with every task completed, no fanin mismatch, no
resource leak and no assertion failure.

+++ esl_proxy/include/algorithm/conf.h
@@ -71,9 +71,13 @@
 #error "EXEC_DURATION_SCALE must be > 0"
 #endif
 
+/*
+ * 返回 uint32_t：raw_duration 可超过 65535（qwen3 的 GATE/UP/DOWN_PROJ 就是
+ * 7-9 万 cycles），若在此截断，EXEC_DURATION_SCALE 较小时缩放结果会回绕。
+ */
 #define SCALE_EXEC_DURATION(raw_duration)                                         \
     (((uint32_t)(raw_duration) > (uint32_t)(EXEC_DURATION_SCALE))               \
-         ? (uint16_t)((uint32_t)(raw_duration) / (uint32_t)(EXEC_DURATION_SCALE)) \
-         : (uint16_t)1)
+         ? (uint32_t)((uint32_t)(raw_duration) / (uint32_t)(EXEC_DURATION_SCALE)) \
+         : (uint32_t)1)
 
 #endif /* ALGORITHM_CONF_H */
