+++ esl_proxy/include/algorithm/task.h
@@ -61,7 +61,7 @@ struct task_desc {
     int64_t        scalar[32];  /* scalar kernel arguments */
     uint16_t       tensor_cnt;  /* number of valid data[] entries */
     uint16_t       scalar_cnt;  /* number of valid scalar[] entries */
-    uint16_t       duration;    /* estimated kernel cycles (low 16 bits) */
+    uint32_t       duration;    /* estimated kernel cycles */
 };
 
 struct predecessor_list {
